feat: 네트워크 모듈 설계
